Replace `DetachedPwmPin` with new `Channel` struct
This is an intermediate step. The plan is to replace `CTimerPwmPin` too.
